requirements.txt
Flask
python-dotenv
gunicorn
sib-api-v3-sdk